requirements-rdkit.txt
im-pipelines-utils==2.4.*
im-pipelines-utils-rdkit==1.5.*
matplotlib==2.2.*
molvs==0.1.1
standardiser==0.1.9
pandas==1.0.1
scikit-learn==0.22.1
